server/requirements.txt
Flask>=3.0.0,<4.0.0
Flask-Cors>=4.0.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
redis>=5.0.0,<6.0.0
weaviate-client>=3.26.0,<5.0.0
psycopg2-binary>=2.9.0,<3.0.0
